requirements.txt
cryptography
tkinter
Pillow==2.2.2